+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Work title and authors info only show up in edition endpoint now
 - Editions and utils tests updated and fixed
 - All toc object url links in ProjectMuse manifests have a query parameter now
+- Deleted extra urls in toc object of Muse manifests 
 
 ## 2022-05-19 -- v0.10.3
 ### Added